requirements.txt
Flask
psycopg2-binary
SQLAlchemy
line-bot-sdk
openai